fix: adjustments to how mykiva page is loaded
